tests/slicing: fix warnings about incompatible pointer types

+++ tests/slicing/sources/ptrtoint6.c
@@ -1,9 +1,11 @@
+#include <stdint.h>
+
 void test_assert(int);
 
-void set(unsigned long addr) { *((int *) addr) = 13; }
+void set(uintptr_t addr) { *((int *) addr) = 13; }
 
 int main(void) {
     int a = 0, b = 0;
-    set(&a);
+    set((uintptr_t) &a);
     test_assert(a == 13);
 }
